yeneta_backend/requirements.txt
# ============================================
# YENETA AI SCHOOL - PYTHON DEPENDENCIES
# ============================================

# --------------------------------------------
# Core Django Packages
# --------------------------------------------
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0

# --------------------------------------------
# Environment Management
# --------------------------------------------
python-dotenv>=1.0.0

# --------------------------------------------
# LLM Providers - Tier 2 (Subsidized/Free)
# --------------------------------------------
google-generativeai>=0.3.2

# --------------------------------------------
# LLM Providers - Tier 3 (Premium) - Optional
# --------------------------------------------
openai>=1.12.0

# --------------------------------------------
# LLM Providers - Tier 1 (Offline/Free)
# --------------------------------------------
ollama>=0.1.7

# --------------------------------------------
# LLM Orchestration & Routing
# --------------------------------------------
langchain>=0.1.9
langchain-community>=0.0.24
langchain-google-genai>=0.0.9
langchain-openai>=0.0.8

# --------------------------------------------
# Token Counting & Cost Optimization
# --------------------------------------------
tiktoken>=0.6.0

# --------------------------------------------
# Vector Database & RAG
# --------------------------------------------
chromadb>=0.4.22
sentence-transformers>=2.3.1

# --------------------------------------------
# Web Search & SERP
# --------------------------------------------
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0

# --------------------------------------------
# Data Processing & Utilities
# --------------------------------------------
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
python-dateutil>=2.8.2
pytz>=2024.1

# --------------------------------------------
# HTTP & Async Support
# --------------------------------------------
httpx>=0.26.0
aiohttp>=3.9.0

# --------------------------------------------
# Text Processing & Document Loaders
# --------------------------------------------
Pillow>=10.2.0
PyPDF2>=3.0.0
python-docx>=1.1.0
pypdf>=3.17.0
reportlab>=4.0.0
docx2txt>=0.8
unstructured>=0.11.0

# --------------------------------------------
# Monitoring & Logging
# --------------------------------------------
colorlog>=6.8.0